project-management-agentic/requirements.txt
# Core dependencies - Compatible versions
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0
langchain-core>=0.1.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
langchain-anthropic>=0.1.1
langgraph>=0.0.20

# Vector stores and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Database
sqlalchemy>=2.0.23
alembic>=1.13.1

# Document processing
pypdf2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
jinja2>=3.1.2

# Web framework (optional for API)
fastapi>=0.104.1
uvicorn>=0.24.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.7
rich>=13.7.0
loguru>=0.7.2

# Development
pytest>=7.4.3
black>=23.11.0
flake8>=6.1.0
